Fix esternal styles with expressions are not transpiled
The problem was that when using interpolation in external files eg. `div { color: ${colors.foo} }` we replace the interpolation `${colors.foo}` with a placeholder so that it can be parsed as css.

For external files we also check whether the export is a string or rather valid css (we do css detection). The problem was that the placeholder is invalid css eg. `div { color: %foo% } }` (`%foo%` is not a valid value for color) so we skipped the file altogether.

I fixed it by using a fake css custom property as placeholder instead.